Update README.md
In this protocol, the Prover possesses a private key x and wants to prove to the Verifier that they know the value of x without revealing it. The Verifier knows the corresponding public key y = G^x.

The protocol proceeds as follows:

The Prover selects a random number r and computes R = G^r.
The Prover sends R to the Verifier.
The Verifier selects a random challenge value c and sends it to the Prover.
The Prover computes s = r + cx and sends s to the Verifier.
The Verifier checks whether R = G^s - y^c. If this equation holds, the Verifier accepts the proof and the Prover is deemed to have proved knowledge of the secret x without revealing it.